Fix pr #3: Fix issue #2: Task 1: Multi-Agent Framework
