+++ manus-open/agents/coordinator_agent.py
@@ -1,0 +1,12 @@
+class CoordinatorAgent:
+    def __init__(self):
+        pass
+
+    def handle_request(self, request):
+        pass
+
+    def store_session(self, session):
+        pass
+
+    def respond_to_user(self, response):
+        pass
