--- manus-open/agents/executor.py
@@ -1,8 +1,0 @@
-class ExecutorAgent:
-    def __init__(self):
-        self.procedural_memory = ProceduralMemory()
-        self.tool_registry = ToolRegistry()
-
-    def execute_plan(self, plan):
-        # Simple execution logic: return the plan as the result
-        return plan
