+++ manus-open/agents/executor_agent.py
@@ -1,5 +1,8 @@
-class ProceduralMemory:
+class ExecutorAgent:
     def __init__(self):
+        pass
+
+    def execute_task(self, task):
         pass
 
     def record_workflow(self, workflow):
