--- manus-open/agents/knowledge.py
@@ -1,7 +1,0 @@
-class KnowledgeAgent:
-    def __init__(self):
-        self.semantic_memory = SemanticMemory()
-
-    def query_knowledge(self, request):
-        # Simple knowledge query logic: return the request as the knowledge
-        return request
