--- manus-open/app/models/model_router.py
@@ -1,19 +1,0 @@
-from fastapi import FastAPI, Request
-from fastapi.responses import JSONResponse
-from pydantic import BaseModel
-from typing import List, Dict
-
-app = FastAPI()
-
-class ModelRequest(BaseModel):
-    task: str
-    data: Dict
-
-@app.post("/route")
-async def route_model(request: Request, model_request: ModelRequest):
-    # Implement model routing logic here
-    return JSONResponse(content={"message": "Model routed successfully"})
-
-if __name__ == "__main__":
-    import uvicorn
-    uvicorn.run(app, host="0.0.0.0", port=8000)
