--- manus-open/app/models/model_server.py
@@ -1,19 +1,0 @@
-from fastapi import FastAPI, Request
-from fastapi.responses import JSONResponse
-from pydantic import BaseModel
-from typing import List, Dict
-
-app = FastAPI()
-
-class InferenceRequest(BaseModel):
-    model: str
-    input_data: Dict
-
-@app.post("/infer")
-async def infer(request: Request, inference_request: InferenceRequest):
-    # Implement inference logic here
-    return JSONResponse(content={"message": "Inference completed successfully"})
-
-if __name__ == "__main__":
-    import uvicorn
-    uvicorn.run(app, host="0.0.0.0", port=8001)
