--- manus-open/core/planner.py
@@ -1,22 +1,0 @@
-class PlannerAgent:
-    def __init__(self, knowledge_agent, executor_agent, procedural_memory):
-        self.knowledge_agent = knowledge_agent
-        self.executor_agent = executor_agent
-        self.procedural_memory = procedural_memory
-
-    def handle_task(self, task, session_id):
-        # Query knowledge agent for relevant information
-        knowledge = self.knowledge_agent.query_knowledge(task)
-
-        # Plan the task based on the knowledge
-        plan = self.plan_task(task, knowledge)
-
-        # Store the plan in procedural memory
-        self.procedural_memory.store_plan(session_id, plan)
-
-        # Execute the plan
-        self.executor_agent.execute_plan(plan, session_id)
-
-    def plan_task(self, task, knowledge):
-        # Implement task planning logic here
-        pass
