+++ manus-open/memory/episodic_memory.py
@@ -1,11 +1,9 @@
 class EpisodicMemory:
     def __init__(self):
-        self.sessions = {}
+        pass
 
-    def store_session(self, request):
-        session_id = len(self.sessions) + 1
-        self.sessions[session_id] = request
-        return session_id
+    def store_session(self, session):
+        pass
 
     def retrieve_session(self, session_id):
-        return self.sessions.get(session_id)
+        pass
