+++ manus-open/models/task_specific_models.py
@@ -1,4 +1,6 @@
 class TaskSpecificModels:
-    def execute(self, task):
-        # Simple task-specific model execution logic: return the task as the result
-        return task
+    def __init__(self):
+        pass
+
+    def execute_specialized_task(self, task):
+        pass
