+++ manus-open/tools/browser_automation.py
@@ -1,11 +1,6 @@
-from selenium import webdriver
-
 class BrowserAutomation:
     def __init__(self):
-        self.driver = webdriver.Chrome()
+        pass
 
-    def execute(self, params):
-        # Implement browser automation logic here
-        self.driver.get(params['url'])
-        # Add more automation steps as needed
-        return self.driver.page_source
+    def automate_browser(self, task):
+        pass
