+++ manus-open/tools/data_processing.py
@@ -1,8 +1,6 @@
-import pandas as pd
-
 class DataProcessing:
-    def execute(self, params):
-        # Implement data processing logic here
-        df = pd.DataFrame(params['data'])
-        # Add more processing steps as needed
-        return df.to_dict()
+    def __init__(self):
+        pass
+
+    def process_data(self, data):
+        pass
