+++ manus-open/tools/tool_registry.py
@@ -1,9 +1,9 @@
 class ToolRegistry:
     def __init__(self):
-        self.tools = {}
+        pass
 
-    def register_tool(self, name, tool):
-        self.tools[name] = tool
+    def register_tool(self, tool):
+        pass
 
-    def get_tool(self, name):
-        return self.tools.get(name)
+    def get_tool(self, tool_name):
+        pass
